Predicting the price of a property in Real Estate
Price of a property is one of the most important decision criterion when people buy homes. Real state firms need to be consistent in their pricing in order to attract buyers . Having a predictive model for the same will be great tool to have , which in turn can also be used to tweak development of properties , putting more emphasis on qualities which increase the value of the property.
We are given you two datasets , housing_train.csv and housing_test.csv . We need to use data housing_train to build predictive model for response variable "Price". Housing_test data contains all other factors except "Price", we need to predict that using the model that you developed and submit our predicted values in a csv files.

Evaluation Criterion : 
Score will be calculated as:

Score =212467/RMSE
Where RMSE is root mean square error on test file